change Angle type

+++ software/test_dir/teensy_test/lib/servos/servos.h
@@ -10,14 +10,14 @@ class Servos{
     public:
         Servos(Adafruit_PWMServoDriver &drv, float freq);
         void begin();
-        void setAngle(byte angle[12]); // update pos[12] variable
-        byte* getAngle(); // return a pointer to an array use getAngle()[i] to return the ième pos
+        void setAngle(int angle[12]); // update pos[12] variable
+        int* getAngle(); // return a pointer to an array use getAngle()[i] to return the ième pos
         void applyChanges(); // update the position directly on the servo
             
     private:
         Adafruit_PWMServoDriver *driver;
 
-        byte pos[12];
+        int pos[12];
         float frequency;
 
 };
